Refactor: getOrderVector를 orderQueryService로 이동
