+++ commerce/src/main/java/store/_0982/commerce/application/order/OrderService.java
@@ -5,20 +5,12 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import store._0982.commerce.application.order.dto.*;
-import store._0982.commerce.application.product.dto.OrderVectorInfo;
-import store._0982.commerce.domain.grouppurchase.GroupPurchase;
-import store._0982.commerce.domain.grouppurchase.GroupPurchaseRepository;
 import store._0982.commerce.domain.order.Order;
-import store._0982.commerce.domain.product.ProductVector;
-import store._0982.commerce.infrastructure.product.ProductVectorJpaRepository;
 import store._0982.common.dto.PageResponse;
 
 import java.util.List;
-import java.util.Map;
 import java.util.UUID;
-import java.util.function.Function;
 
-import static java.util.stream.Collectors.toMap;
 @Slf4j
 @Service
 @RequiredArgsConstructor
@@ -26,9 +18,6 @@ public class OrderService {
 
     private final OrderCommandService orderCommandService;
     private final OrderQueryService orderQueryService;
-    private final GroupPurchaseRepository groupPurchaseRepository;
-    private final ProductVectorJpaRepository productVectorRepository;
-
     /**
      * 주문 생성
      *
@@ -101,43 +90,5 @@ public class OrderService {
      */
     public List<Order> getAllOrderByMemberId(UUID memberId) {
         return orderQueryService.getAllOrderByMemberId(memberId);
-    }
-
-    /**
-     * internal orderVector 조회
-     *
-     * @param memberId
-     * @return List<OrderVectorInfo>
-     */
-    public List<OrderVectorInfo> getOrderVector(UUID memberId) {
-        List<Order> orders = orderQueryService.getAllOrderByMemberId(memberId);
-        List<UUID> groupPurchaseIds = orders.stream()
-                .map(Order::getGroupPurchaseId)
-                .toList();
-        List<GroupPurchase> groupPurchases = groupPurchaseRepository.findAllByGroupPurchaseIdIn(groupPurchaseIds);
-        List<UUID> productIds = groupPurchases.stream()
-                .map(GroupPurchase::getProductId)
-                .toList();
-        List<ProductVector> productVectors = productVectorRepository.findByProductIdIn(productIds);
-        Map<UUID, UUID> groupPurchaseToProduct = groupPurchases.stream()
-                .collect(toMap(GroupPurchase::getGroupPurchaseId, GroupPurchase::getProductId));
-        Map<UUID, ProductVector> productIdToVector = productVectors.stream()
-                .collect(toMap(ProductVector::getProductId, Function.identity()));
-        return orders.stream()
-                .map(order -> {
-                    UUID productId = groupPurchaseToProduct.get(order.getGroupPurchaseId());
-                    ProductVector vector = productIdToVector.get(productId);
-                    float[] productVector = vector == null ? null : vector.getVector();
-                    return new OrderVectorInfo(
-                            order.getOrderId(),
-                            order.getMemberId(),
-                            productId,
-                            order.getQuantity(),
-                            order.getCreatedAt(),
-                            order.getStatus(),
-                            productVector
-                    );
-                })
-                .toList();
     }
 }
